proyectoFinal/requirements.txt
fastapi
uvicorn
scikit-learn==1.5.1
joblib
numpy
python-multipart
pandas
pymysql
scipy